refactor (components/common): improve code organization and component structure with our common Empty component
