+++ Code/Web/src/components/common/index.ts
@@ -3,3 +3,4 @@ export { default as Input } from './Input/Input'
 export { default as Modal } from './Modal/Modal'
 export { default as Card } from './Card/Card'
 export { default as StatusTag } from './Tag/StatusTag'
+export { default as Empty } from './Empty/Empty'
